feat(diagnostics): wire missing-import detection into live publishDiagnostics
Adds features::missing_import_diagnostics, wired into document_handler.rs
alongside call_arg_diagnostics/nullable_dot_call_diagnostics (both call
sites: didOpen/didChange and republish_open_file_diagnostics). Same
DiagnosticSeverity::WARNING/source="kmp-lsp" convention as the sibling
diagnostics, and the same jar_phase.is_loading() suppression gate as
call_arg_diagnostics (this diagnostic's "importable"/default-import checks
read jar_definitions/jar_qualified directly, so a partially-materialized JAR
index can transiently misjudge either direction; the actor's existing
republish-on-jar-done cycle self-heals it, no new machinery needed).

The candidate-collection walk and per-candidate resolve checks (developed
and precision-tuned against real nowInAndroid/Moneta corpora in the prior
3 commits) are now the single shared implementation
(collect_missing_import_flags), consumed by both the new live diagnostic and
the existing `missing-imports` CLI precision-measurement subcommand, which
is refactored to call the shared function instead of duplicating it.

Live-verified via the real LSP protocol (not just the CLI): opening a real
nowInAndroid file produces zero diagnostics; opening a synthetic file with a
genuine unresolved reference produces exactly one, with the correct range
and message, appearing only once JAR loading settles (confirming the
gate/republish cycle actually round-trips end to end, not just in unit
tests).

Also noted, not fixed: two consecutive missing-imports CLI runs against the
same Moneta checkout (same binary, no code change) produced slightly
different counts (24 vs 25) -- pre-existing nondeterminism somewhere in the
concurrent JAR-indexing pipeline, not something this change introduced.
Worth its own investigation before relying on exact false-positive counts,
though the live diagnostic's per-file, on-demand nature narrows its
practical impact versus the CLI's whole-workspace aggregate.

New tests (features/missing_import_diagnostics_tests.rs): flags a
known-elsewhere unimported type, suppresses for an explicit import,
suppresses for a default-import stdlib type, and suppresses/resumes across
the jar-loading gate (mirroring call_arg_diagnostics's own gate test).
